refactor(file): Move FileNaming to its own file.
